Add Video suggester façade Remove :fire: TermSuggester         Co-authored-by: Cristina <Cristinalpena@gmail.com>         Co-authored-by: Kathleen <khakicarroll@comcast.net>         Co-authored-by: Cassie <caachzenick@gmail.com>
