--- spec/support/fixtures/vcr_cassettes/YoutubeService/call_on_youtube_api.yml
@@ -1,57 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://www.googleapis.com/youtube/v3/search/part=snippet&maxResults=3&q=cute%20puppies&key=AIzaSyDOrUy7uvek6tj4n9xasEgDdB2MQRkmOr8
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v1.0.1
-      X-Honeycomb-Trace:
-      - 1;dataset=,trace_id=03ff957c-060d-434e-b9c7-48eefbdbe873,parent_id=1f70d025-8919-4134-8bcd-b63a233d5ce7,context=e30=
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 404
-      message: Not Found
-    headers:
-      Cache-Control:
-      - no-cache, no-store, max-age=0, must-revalidate
-      Pragma:
-      - no-cache
-      Expires:
-      - Mon, 01 Jan 1990 00:00:00 GMT
-      Date:
-      - Sat, 16 May 2020 17:34:09 GMT
-      Vary:
-      - Origin
-      - X-Origin
-      Content-Type:
-      - text/html; charset=UTF-8
-      X-Content-Type-Options:
-      - nosniff
-      X-Frame-Options:
-      - SAMEORIGIN
-      Content-Security-Policy:
-      - frame-ancestors 'self'
-      X-Xss-Protection:
-      - 1; mode=block
-      Server:
-      - GSE
-      Alt-Svc:
-      - h3-27=":443"; ma=2592000,h3-25=":443"; ma=2592000,h3-T050=":443"; ma=2592000,h3-Q050=":443";
-        ma=2592000,h3-Q049=":443"; ma=2592000,h3-Q048=":443"; ma=2592000,h3-Q046=":443";
-        ma=2592000,h3-Q043=":443"; ma=2592000,quic=":443"; ma=2592000; v="46,43"
-      Transfer-Encoding:
-      - chunked
-    body:
-      encoding: ASCII-8BIT
-      string: Not Found
-    http_version: null
-  recorded_at: Sat, 16 May 2020 17:34:09 GMT
-recorded_with: VCR 5.1.0
